Move most syntax classes out of Process.scala
